Add support for showing GridLines, organized common code
Showing gridlines can be very useful when drawing on the canvas,
so we added a method to show/hide it together with a tooltip
that shows the current location under the mouse (x,y coordinates).

This also prompted a reorg of the code since other locales will
need the same behavior, so we moved that to a shared project
which now also contains the window snapping behavior which
will also be shared.