Removed all Logo namespaces and made a static Logo class instead
This allows providing convenience static methods that look like
part of the language (kinda ;)), like Repetir(....). Still looks
a bit alien with the () => lambda syntax, but better than teaching
for (var.... ) ;)

Also made the Turtle "chainable" so you can invoke multiple statements
in a single train of calls.

Shielded the Show/Hide methods also from an issue in SmallBasic that
causes a WPF exception if Show is called when the Turtle is already
visible.